requirements/test.txt
black==25.1.0
django-debug-toolbar==5.1.0
flake8==7.1.2